third_party/WebKit/LayoutTests/bluetooth/requestDevice/discovery-succeeds.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Discover a device using alias, name, or UUID.';

bluetooth_test(() => {
  let devices = [];
  let push = device => devices.push(device);
  // We can't use Promise.all() because Chrome will always close
  // the previous chooser in the process of handling a user gesture
  // for the next request.
  return getHealthThermometerDevice()
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.alias] }]
    }))
    .then(push)
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.name] }]
    }))
    .then(push)
    .then(() => requestDeviceWithTrustedClick({
      filters: [{ services: [health_thermometer.uuid] }]
    }))
    .then(push)
    devices.forEach(device => {
      assert_equals(device.constructor.name, 'BluetoothDevice');
    });
}, test_desc);
</script>
